refactor: move Music class under view/music from view
